move _actionsHelpers to _activeMetadata decouple actions and constants
