+++ src/state/_activeMetadata/constants.js
@@ -1,8 +1,5 @@
-import _ from 'lodash';
-
-import common from './_common/actions';
-import Select from './Select';
-import ActionTypes from '../constants/ActionTypes';
+import Select from '../Select';
+import ActionTypes from '../../constants/ActionTypes';
 
 export const STORES_TO_ENSURE_WITH_FILTER_BY_ACTIVE = {
 	areaRelations: [
@@ -44,24 +41,3 @@ export const STORES_TO_ENSURE_WITH_FILTER_BY_ACTIVE = {
 	tags: [Select.tags.getSubstate, 'tags', ActionTypes.TAGS],
 	views: [Select.views.getSubstate, 'views', ActionTypes.VIEWS, 'views'],
 };
-
-/**
- * Ensure indexes with filter by active for each store from the list (above)
- *
- * @param filterKey {string} active metadata type key (e.g. scope, place, ..)
- */
-export function ensureDependenciesOfActiveMetadataType(filterKey) {
-	return dispatch => {
-		let filterByActive = {
-			[filterKey]: true,
-		};
-
-		_.forIn(STORES_TO_ENSURE_WITH_FILTER_BY_ACTIVE, params => {
-			dispatch(
-				common.ensureIndexesWithFilterByActive(...params)(filterByActive)
-			);
-		});
-
-		// TODO data-based stores
-	};
-}
